Agregué la accion de comprar el carrito de compras, tambien meter dinero falso llenar un campo de la base de datos y las vistas de comprados, seguimiento de pedido y de vendedor

+++ ECommerce/src/main/java/Modelo/Usuario.java
@@ -15,7 +15,7 @@ public class Usuario {
     private Date Fecha_Registro;
     private Date Fecha_Nacimiento;
     private String ImagenURL;
-    private float DineroFalso;
+    private double DineroFalso;
 
    
     // Constructor
@@ -137,11 +137,11 @@ public class Usuario {
         this.ImagenURL = ImagenURL;
     }
     
-    public float getDineroFalso() {
+    public double getDineroFalso() {
         return DineroFalso;
     }
 
-    public void setDineroFalso(float DineroFalso) {
+    public void setDineroFalso(double DineroFalso) {
         this.DineroFalso = DineroFalso;
     }
 }
